requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
PyPDF2
pytesseract
pillow
python-dotenv